requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.4.2
transformers==4.40.2
accelerate==0.29.3
torch==2.3.0
google-cloud-aiplatform>=1.36.0
vertexai>=0.0.1
google-cloud-firestore>=2.0.0 